requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
coreapi>=2.3.3
markdown>=3.5.1
django-filter>=23.4.0
psycopg2-binary>=2.9.9  # Pour PostgreSQL
python-dotenv>=1.0.0    # Pour les variables d'environnement
